+++ app/src/main/java/nl/frankkie/convention/EventListActivity.java
@@ -203,12 +203,12 @@ public class EventListActivity extends ActionBarActivity
             }
         } else {
             //GCM is available!!
-            String regId = Util.gcmGetRegId(this);
+            String regId = GcmUtil.gcmGetRegId(this);
             //not using String.isEmpty, as this might not work on some older Android versions.
             //not sure from which version isEmpty is supported, not taking any chances here.
             if (regId == null || "".equals(regId)){
                 //not registered yet
-                Util.gcmRegister(this);
+                GcmUtil.gcmRegister(this);
             }
         }
     }
